+++ CHANGELOG.md
@@ -12,6 +12,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Jitsi streamed in marsha live
 
+### Fixed
+
+- Fetch fresh resource data after initiate-upload endpoint called
+
 ## [3.18.0] - 2021-05-10
 
 ### Added